requirements.txt
aiocache==0.12.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==7.1.2
elasticsearch==7.9.1
fastapi==0.61.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.4
orjson==3.8.7
psycopg2==2.9.5
pydantic==1.9.0
python-dotenv==0.21.1
redis==4.4.2
starlette==0.13.6
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.12.2
uvloop==0.17.0
yarl==1.8.2
PyJWT==2.6.0
requests==2.28.2

pip~=22.3.1
wheel~=0.38.4
setuptools~=65.5.1
Django~=4.2.2
asgiref~=3.7.2
sqlparse~=0.4.4